% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowplyr-package.R
\docType{package}
\name{flowplyr-package}
\alias{flowplyr}
\alias{flowplyr-package}
\title{flowplyr: A modern, flexible toolkit for flow cytometry data in R.}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hugoakerstrand.github.io/flowplyr/}
}

}
\author{
\strong{Maintainer}: Hugo Åkerstrand \email{hugo.akerstrand@gmail.com}

}
\keyword{internal}
