% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tibble.R
\name{to_tibble}
\alias{to_tibble}
\title{Convert a flowSet into a tibble}
\usage{
to_tibble(data, ...)
}
\arguments{
\item{data}{A flowSet object.}

\item{...}{What slots to extract? <kw: default = "all"> //  One or more of
\code{"all"}, \code{"sample_names"}, \code{"exprs"}, \code{"keywords"}, or \code{"meta_data"}.}
}
\value{
A tibble containing the requested data components. The function will
error if \code{data} is not a flowSet or if invalid keywords are provided.
}
\description{
\code{to_tibble()} creates a \code{\link[tibble:tibble]{tibble::tibble()}} from a
\code{\link[flowCore:flowSet-class]{flowCore::flowSet()}}; an S4 class that stores .fcs file data.
}
